Moved component into parts
