+++ src/parts/MovieCarousel.tsx
